(PC-39170)[API] refactor: separate venue serialize in chunks
