+++ api/src/pcapi/routes/serialization/venue_finance_serialize.py
@@ -2,6 +2,28 @@ import enum
 
 from pcapi.core.finance import models as finance_models
 from pcapi.core.offerers import models as offerers_models
+from pcapi.routes.serialization import BaseModel
+
+
+class LinkVenueToPricingPointBodyModel(BaseModel):
+    pricingPointId: int
+
+    class Config:
+        extra = "forbid"
+
+
+class GetVenuePricingPointResponseModel(BaseModel):
+    id: int
+    siret: str
+    venueName: str
+
+    class Config:
+        orm_mode = True
+
+    @classmethod
+    def from_orm(cls, venue: offerers_models.Venue) -> "GetVenuePricingPointResponseModel":
+        venue.venueName = venue.publicName
+        return super().from_orm(venue)
 
 
 class SimplifiedBankAccountStatus(enum.Enum):
