+++ api/src/pcapi/routes/serialization/venue_serialize.py
@@ -14,8 +14,6 @@ from pydantic.v1 import validator
 
 from pcapi.connectors.serialization import acceslibre_serializers
 from pcapi.core.educational import models as educational_models
-from pcapi.core.educational.constants import ALL_INTERVENTION_AREA
-from pcapi.core.finance import models as finance_models
 from pcapi.core.geography import utils as geography_utils
 from pcapi.core.geography.constants import MAX_LATITUDE
 from pcapi.core.geography.constants import MAX_LONGITUDE
@@ -23,8 +21,6 @@ from pcapi.core.offerers import exceptions
 from pcapi.core.offerers import models as offerers_models
 from pcapi.core.offerers import schemas as offerers_schemas
 from pcapi.core.offerers.constants import JE_VEUX_AIDER_GOUV_BASE_URL
-from pcapi.core.offerers.validation import VENUE_BANNER_MAX_SIZE
-from pcapi.core.offers.validation import ACCEPTED_THUMBNAIL_FORMATS
 from pcapi.core.opening_hours import api as opening_hours_api
 from pcapi.core.opening_hours import schemas as opening_hours_schemas
 from pcapi.routes.native.v1.serialization.common_models import AccessibilityComplianceMixin
@@ -32,49 +28,16 @@ from pcapi.routes.serialization import BaseModel
 from pcapi.routes.serialization import HttpBodyModel
 from pcapi.routes.serialization import address_serialize
 from pcapi.routes.serialization.venue_types_serialize import VenueTypeResponseModel
-from pcapi.routes.shared.collective.serialization import offers as shared_offers
-from pcapi.serialization.utils import string_length_validator
 from pcapi.serialization.utils import string_to_boolean_field
 from pcapi.serialization.utils import to_camel
 from pcapi.utils import date as date_utils
 from pcapi.utils.date import format_into_utc_date
-from pcapi.utils.image_conversion import CropParam
-from pcapi.utils.image_conversion import CropParams
+
+from . import venue_banners_serialize
+from . import venue_collective_serialize
+from . import venue_finance_serialize
 
 from . import deprecated  # noqa: F401
-
-
-class DMSApplicationstatus(enum.Enum):
-    ACCEPTED = "accepte"
-    DROPPED = "sans_suite"
-    BUILDING = "en_construction"
-    REFUSED = "refuse"
-    INSTRUCTING = "en_instruction"
-
-
-class DMSApplicationForEAC(BaseModel):
-    venueId: int
-    state: DMSApplicationstatus
-    procedure: int
-    application: int
-    lastChangeDate: datetime
-    depositDate: datetime
-    expirationDate: datetime | None
-    buildDate: datetime | None
-    instructionDate: datetime | None
-    processingDate: datetime | None
-    userDeletionDate: datetime | None
-
-    class Config:
-        orm_mode = True
-        json_encoders = {datetime: format_into_utc_date}
-
-    @classmethod
-    def from_orm(  # type: ignore[override]
-        cls, collective_dms_application: educational_models.CollectiveDmsApplication, venue_id: int
-    ) -> "DMSApplicationForEAC":
-        collective_dms_application.venueId = venue_id  # type: ignore [attr-defined]
-        return super().from_orm(collective_dms_application)
 
 
 class PostVenueBodyModel(BaseModel, AccessibilityComplianceMixin):
@@ -126,80 +89,6 @@ class GetVenueManagingOffererResponseModel(BaseModel):
         json_encoders = {datetime: format_into_utc_date}
 
 
-class BannerMetaModel(BaseModel):
-    image_credit: offerers_schemas.VenueImageCredit | None = None
-    original_image_url: str | None = None
-    crop_params: CropParams = CropParams()
-
-    @validator("crop_params", pre=True)
-    @classmethod
-    def validate_crop_params(cls, raw_crop_params: CropParams | None) -> CropParams:
-        """
-        Old venues might have a crop_params key with a null value
-        """
-        if not raw_crop_params:
-            return CropParams()
-        return raw_crop_params
-
-
-class GetVenuePricingPointResponseModel(BaseModel):
-    id: int
-    siret: str
-    venueName: str
-
-    class Config:
-        orm_mode = True
-
-    @classmethod
-    def from_orm(cls, venue: offerers_models.Venue) -> "GetVenuePricingPointResponseModel":
-        venue.venueName = venue.publicName  # type: ignore [attr-defined]
-        return super().from_orm(venue)
-
-
-class GetVenueDomainResponseModel(BaseModel):
-    id: int
-    name: str
-
-    class Config:
-        orm_mode = True
-
-
-class LegalStatusResponseModel(BaseModel):
-    id: int
-    name: str
-
-    class Config:
-        orm_mode = True
-
-
-class SimplifiedBankAccountStatus(enum.Enum):
-    PENDING = "pending"
-    VALID = "valid"
-    PENDING_CORRECTIONS = "pending_corrections"
-
-
-def parse_venue_bank_account_status(venue: offerers_models.Venue) -> SimplifiedBankAccountStatus | None:
-    status_enum = finance_models.BankAccountApplicationStatus
-    bank_account = venue.current_bank_account
-
-    # TODO(jbaudet - 10/2025): move this code to a more
-    # appropriate api/repository/models module once offerers
-    # have been replaced by venues. These rules come from the
-    # `get_offerer_and_extradata` function.
-    if not bank_account or not bank_account.isActive:
-        return None
-
-    match bank_account.status:
-        case status_enum.ACCEPTED:
-            return SimplifiedBankAccountStatus.VALID
-        case status_enum.DRAFT | status_enum.ON_GOING:
-            return SimplifiedBankAccountStatus.PENDING
-        case status_enum.WITH_PENDING_CORRECTIONS:
-            return SimplifiedBankAccountStatus.PENDING_CORRECTIONS
-        case _:
-            return None
-
-
 class GetVenueResponseGetterDict(pydantic_v1.utils.GetterDict):
     def get(self, key: str, default: typing.Any | None = None) -> typing.Any:
         venue: offerers_models.Venue = self._obj
@@ -248,7 +137,7 @@ class GetVenueResponseGetterDict(pydantic_v1.utils.GetterDict):
 
         if key == "collectiveDmsApplications":
             return [
-                DMSApplicationForEAC.from_orm(collective_ds_application, venue.id)
+                venue_collective_serialize.DMSApplicationForEAC.from_orm(collective_ds_application, venue.id)
                 for collective_ds_application in venue.collectiveDmsApplications
             ]
 
@@ -276,7 +165,7 @@ class GetVenueResponseGetterDict(pydantic_v1.utils.GetterDict):
             return venue.managingOfferer.allowedOnAdage
 
         if key == "bankAccountStatus":
-            return parse_venue_bank_account_status(venue)
+            return venue_finance_serialize.parse_venue_bank_account_status(venue)
 
         if key == "hasNonFreeOffers":
             # avoid some tricky circular import: schemas is not expected
@@ -321,24 +210,24 @@ class GetVenueResponseModel(BaseModel, AccessibilityComplianceMixin):
     activity: offerers_models.DisplayableActivity | None
     dateCreated: datetime
     id: int
-    bannerMeta: BannerMetaModel | None
+    bannerMeta: venue_banners_serialize.BannerMetaModel | None
     bookingEmail: str | None
     comment: str | None
     managingOfferer: GetVenueManagingOffererResponseModel
-    pricingPoint: GetVenuePricingPointResponseModel | None
+    pricingPoint: venue_finance_serialize.GetVenuePricingPointResponseModel | None
     siret: str | None
     venueType: VenueTypeResponseModel
     collectiveDescription: str | None
     collectiveStudents: list[educational_models.StudentLevels] | None
     collectiveWebsite: str | None
-    collectiveDomains: list[GetVenueDomainResponseModel]
+    collectiveDomains: list[venue_collective_serialize.GetVenueDomainResponseModel]
     collectiveInterventionArea: list[str] | None
-    collectiveLegalStatus: LegalStatusResponseModel | None
+    collectiveLegalStatus: venue_collective_serialize.LegalStatusResponseModel | None
     collectiveNetwork: list[str] | None
     collectiveAccessInformation: str | None
     collectivePhone: str | None
     collectiveEmail: str | None
-    collectiveDmsApplications: list[DMSApplicationForEAC]
+    collectiveDmsApplications: list[venue_collective_serialize.DMSApplicationForEAC]
     hasAdageId: bool
     adageInscriptionDate: datetime | None
     hasOffers: bool
@@ -349,7 +238,7 @@ class GetVenueResponseModel(BaseModel, AccessibilityComplianceMixin):
     isActive: bool
     isValidated: bool
     allowedOnAdage: bool
-    bankAccountStatus: SimplifiedBankAccountStatus | None
+    bankAccountStatus: venue_finance_serialize.SimplifiedBankAccountStatus | None
     hasNonFreeOffers: bool
     hasPartnerPage: bool
     canDisplayHighlights: bool
@@ -363,7 +252,9 @@ class GetVenueResponseModel(BaseModel, AccessibilityComplianceMixin):
 
     @validator("bannerMeta")
     @classmethod
-    def validate_banner_meta(cls, meta: BannerMetaModel | None, values: dict) -> BannerMetaModel | None:
+    def validate_banner_meta(
+        cls, meta: venue_banners_serialize.BannerMetaModel | None, values: dict
+    ) -> venue_banners_serialize.BannerMetaModel | None:
         """
         Old venues might have a banner url without banner meta, or an
         incomplete banner meta.
@@ -373,32 +264,9 @@ class GetVenueResponseModel(BaseModel, AccessibilityComplianceMixin):
             return None
 
         if not meta:
-            return BannerMetaModel()
+            return venue_banners_serialize.BannerMetaModel()
 
         return meta
-
-
-class GetCollectiveVenueResponseModel(BaseModel):
-    collectiveDescription: str | None
-    collectiveStudents: list[educational_models.StudentLevels] | None
-    collectiveWebsite: str | None
-    collectiveDomains: list[GetVenueDomainResponseModel]
-    collectiveInterventionArea: list[str] | None
-    collectiveLegalStatus: LegalStatusResponseModel | None
-    collectiveNetwork: list[str] | None
-    collectiveAccessInformation: str | None
-    collectivePhone: str | None
-    collectiveEmail: str | None
-    siret: str | None
-
-    class Config:
-        orm_mode = True
-        json_encoders = {datetime: format_into_utc_date}
-
-    @classmethod
-    def from_orm(cls, venue: offerers_models.Venue) -> "GetCollectiveVenueResponseModel":
-        venue.collectiveLegalStatus = venue.venueEducationalStatus  # type: ignore [attr-defined]
-        return super().from_orm(venue)
 
 
 class EditVenueBodyModel(BaseModel, AccessibilityComplianceMixin):
@@ -487,39 +355,6 @@ class EditVenueBodyModel(BaseModel, AccessibilityComplianceMixin):
         return volunteering_url
 
 
-class EditVenueCollectiveDataBodyModel(BaseModel):
-    collectiveDescription: str | None
-    collectiveStudents: list[educational_models.StudentLevels] | None
-    collectiveWebsite: str | None
-    collectiveDomains: list[int] | None
-    collectiveInterventionArea: list[str] | None
-    venueEducationalStatusId: int | None
-    collectiveNetwork: list[str] | None
-    collectiveAccessInformation: str | None
-    collectivePhone: str | None
-    collectiveEmail: str | None
-    activity: offerers_models.ActivityOpenToPublic | offerers_models.ActivityNotOpenToPublic | None
-
-    _validate_collectiveDescription = string_length_validator("collectiveDescription", length=500)
-    _validate_collectiveWebsite = string_length_validator("collectiveWebsite", length=150)
-    _validate_collectiveAccessInformation = string_length_validator("collectiveAccessInformation", length=500)
-    _validate_collectivePhone = string_length_validator("collectivePhone", length=50)
-    _validate_collectiveEmail = string_length_validator("collectiveEmail", length=150)
-
-    @validator("collectiveStudents")
-    def validate_students(cls, students: list[str]) -> list[educational_models.StudentLevels] | None:
-        if not students:
-            return []
-        return shared_offers.validate_students(students)
-
-    @validator("collectiveInterventionArea")
-    def validate_intervention_area(cls, intervention_area: list[str] | None) -> list[str] | None:
-        if intervention_area and any(area not in ALL_INTERVENTION_AREA for area in intervention_area):
-            raise ValueError("One or more element is not a valid area")
-
-        return intervention_area
-
-
 class VenueListItemLiteResponseModel(BaseModel):
     id: int
     name: str
@@ -540,109 +375,6 @@ class VenueListQueryModel(BaseModel):
     class Config:
         alias_generator = to_camel
         extra = "forbid"
-
-
-class VenueBannerContentModel(BaseModel):
-    if typing.TYPE_CHECKING:  # https://github.com/pydantic/pydantic/issues/156
-        content: bytes
-    else:
-        content: pydantic_v1.conbytes(min_length=2, max_length=VENUE_BANNER_MAX_SIZE)
-    image_credit: offerers_schemas.VenueImageCredit | None
-
-    # cropping parameters must be a % (between 0 and 1) of the original
-    # bottom right corner and the original height
-    x_crop_percent: CropParam
-    y_crop_percent: CropParam
-    height_crop_percent: CropParam
-    width_crop_percent: CropParam
-
-    class Config:
-        extra = pydantic_v1.Extra.forbid
-        anystr_strip_whitespace = True
-
-    @root_validator(pre=True)
-    @classmethod
-    def validate_banner(cls, values: dict) -> dict:
-        """
-        Validate content (is not an invalid image) using PIL
-        + set and validate content type using image build from previous
-          step
-        """
-        try:
-            image = Image.open(BytesIO(values["content"]))
-        except Exception:
-            raise ValueError("Format de l'image invalide")
-        assert image.format is not None  # helps mypy
-
-        content_type = image.format.lower()
-
-        if content_type not in ACCEPTED_THUMBNAIL_FORMATS:
-            raise ValueError("Format de l'image invalide")
-
-        return values
-
-    @classmethod
-    def from_request(cls, request: typing.Any) -> "VenueBannerContentModel":
-        cls.validate_request(request)
-
-        file = request.files["banner"]
-        return VenueBannerContentModel(
-            content=file.read(VENUE_BANNER_MAX_SIZE),
-            image_credit=request.args.get("image_credit"),
-            x_crop_percent=request.args.get("x_crop_percent"),
-            y_crop_percent=request.args.get("y_crop_percent"),
-            height_crop_percent=request.args.get("height_crop_percent"),
-            width_crop_percent=request.args.get("width_crop_percent"),
-        )
-
-    @classmethod
-    def validate_request(cls, request: typing.Any) -> typing.Any:
-        """
-        If the request has a content_lenght information, use directly to
-        avoid reading the whole content to check its size. If not, do not
-        consider this a an error: it will be checked later.
-        """
-        try:
-            file = request.files["banner"]
-        except (AttributeError, KeyError):
-            raise exceptions.InvalidVenueBannerContent("Image manquante")
-
-        if file.content_length and file.content_length > VENUE_BANNER_MAX_SIZE:
-            raise exceptions.VenueBannerTooBig(f"Image trop grande, max: {VENUE_BANNER_MAX_SIZE / 1_000}Ko")
-
-        return request
-
-    @property
-    def crop_params(self) -> CropParams | None:
-        if {self.x_crop_percent, self.y_crop_percent, self.height_crop_percent, self.width_crop_percent} == {None}:
-            return None
-
-        return CropParams(
-            x_crop_percent=self.x_crop_percent,
-            y_crop_percent=self.y_crop_percent,
-            height_crop_percent=self.height_crop_percent,
-            width_crop_percent=self.width_crop_percent,
-        )
-
-
-class LinkVenueToPricingPointBodyModel(BaseModel):
-    pricingPointId: int
-
-    class Config:
-        extra = "forbid"
-
-
-class VenuesEducationalStatusResponseModel(BaseModel):
-    id: int
-    name: str
-
-    class Config:
-        orm_mode = True
-        extra = pydantic_v1.Extra.forbid
-
-
-class VenuesEducationalStatusesResponseModel(BaseModel):
-    statuses: list[VenuesEducationalStatusResponseModel]
 
 
 class GetOffersStatsResponseModel(HttpBodyModel):
